refactor products page
